% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDef.R, R/zzz.R
\docType{class}
\name{MiRBaseVersionsDb-class}
\alias{.MiRBaseVersionsDb}
\alias{.onLoad}
\alias{MiRBaseVersionsDb}
\alias{MiRBaseVersionsDb-class}
\alias{columns}
\alias{columns,MiRBaseVersionsDb-method}
\alias{keys}
\alias{keys,MiRBaseVersionsDb-method}
\alias{keytypes}
\alias{keytypes,MiRBaseVersionsDb-method}
\alias{select}
\alias{select,MiRBaseVersionsDb-method}
\title{Database class}
\usage{
columns(x)
keytypes(x)
keys(x, keytype)
select(x, keys, columns, keytype)

\S4method{columns}{MiRBaseVersionsDb}(x)

\S4method{keytypes}{MiRBaseVersionsDb}(x)

\S4method{keys}{MiRBaseVersionsDb}(x, keytype)

\S4method{select}{MiRBaseVersionsDb}(x, keys, columns = "*", keytype = "")

columns(x)
keytypes(x)
keys(x, keytype)
select(x, keys, columns, keytype)
}
\arguments{
\item{x}{the \code{MiRBaseVersionsDb} object}

\item{keytype}{represents the table from which data shall be received. All
possible keytypes can be viewed by using the \code{keytypes} method.}

\item{keys}{the accession name of mature miRNAs. All possible keys (miRNAs)
are returned by using the \code{keys} method.}

\item{columns}{that can be returned for each miRNA. All possible columns can
shown by using the \code{columns} method.}

\item{x}{the \code{MiRBaseVersionsDb} object}

\item{keytype}{represents the table from which data shall be received. All
possible keytypes can be viewed by using the \code{keytypes} method.}

\item{keys}{the accession name of mature miRNAs. All possible keys (miRNAs)
are returned by using the \code{keys} method.}

\item{columns}{that can be returned for each miRNA. All possible columns can
shown by using the \code{columns} method.}

\item{MiRBaseVersionsDb}{object reference}

\item{MiRBaseVersionsDb}{object reference}

\item{MiRBaseVersionsDb}{object reference}

\item{MiRBaseVersionsDb}{object reference}
}
\value{
string vectors, for \code{select} a data.frame with selected columns.

string vectors, for \code{select} a data.frame with selected columns.

Character vector with lower case column names.

Character vector containing upper case column names

Character vector where the values are lower name and the 'name'
attribute are upper name table names.

Character vector with accession names from specified \code{keytype}
}
\description{
object of \code{MiRBaseVersionsDb} class holds the sqlite
database connection, and extends \code{AnnotationDb} class from AnnotationDbi
package. \code{columns}, \code{keys}, \code{keytypes} and \code{select}
methods allow access to database tables and retrieval of miRNA target
information.

\code{select} is for querying the database to retrieve information about
mature miRNA names from selected versions.

object of \code{MiRBaseVersionsDb} class holds the sqlite
database connection, and extends \code{AnnotationDb} class from AnnotationDbi
package. \code{columns}, \code{keys}, \code{keytypes} and \code{select}
methods allow access to database tables and retrieval of miRNA target
information.

\code{select} is for querying the database to retrieve information about
mature miRNA names from selected versions.
}
\details{
This function reads column names from the view of the
current version in lower case. All visible keytypes have the same columns.

This function converts lower case column names (received from
\code{\link{.getLCColNames}}) into upper case.

The \code{.getTableNames} function reads table names from the
database. Two types of tables will be returned, the general mimat table and
views respectively for each miRBase version.

The \code{.keys} function implementation reads the accessions of the
desired keytype.
}
\examples{
#first load the annotations
require(miRBaseVersions.db)
#see all available tables
keytypes(miRBaseVersions.db)
#first load the annotations
require(miRBaseVersions.db)
#see all available tables
keytypes(miRBaseVersions.db)
}
\author{
Stefan Haunsberger \email{stefanhaunsberger@rcsi.ie}

Stefan Haunsberger \email{stefanhaunsberger@rcsi.ie}
}

