% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassDef.R, R/zzz.R
\docType{methods}
\name{columns,MiRBaseVersionsDb-method}
\alias{columns,MiRBaseVersionsDb-method}
\alias{keys,MiRBaseVersionsDb-method}
\alias{keytypes,MiRBaseVersionsDb-method}
\alias{miRBaseVersions.db}
\alias{miRBaseVersions.db-package}
\title{miRBaseVersions.db: miRNA name collection of 21 different miRBase
releases.}
\usage{
\S4method{columns}{MiRBaseVersionsDb}(x)

\S4method{keytypes}{MiRBaseVersionsDb}(x)

\S4method{keys}{MiRBaseVersionsDb}(x, keytype)
}
\details{
This annotation package holds mature miRNA names from 21 different
miRBase versions. It contains one main table containing all miRNAs and
one view for each version, such as 'vw-mimat-21.0' for mature miRNA names
from version 21.0.
}
\author{
Stefan Haunsberger \email{stefanhaunsberger@rcsi.ie}
}

