% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{miRBaseVersions.db}
\alias{miRBaseVersions.db}
\alias{miRBaseVersions.db-package}
\title{miRBaseVersions.db: miRNA name collection of 21 different miRBase
release versions.}
\details{
This annotation package holds mature miRNA names from 21 different
miRBase versions. It contains one main table holding all miRNAs and
one view for each version, such as 'vw-mimat-21.0' for mature miRNA names
from version 21.0.
}
\author{
Stefan Haunsberger \email{stefanhaunsberger@rcsi.ie}
}

